Create Reversed Linked List
You are given a linked list that contains  integers. You have performed the following reverse operation on the list:

Select all the subparts of the list that contain only even integers. For example, if the list is , then the selected subparts will be , .
Reverse the selected subpart such as  and .
Now, you are required to retrieve the original list.

Note: You should use the following definition of the linked list for this problem:

class Node {
    Object data;
    Node next;
}
Input format

First line:  
Next line:  space-separated integers that denote elements of the reverse list
Output format

Print the  elements of the original list.